requirements.txt
openai==1.58.1
vllm==0.6.6.post1
gritlm==1.0.2
torch==2.5.1
transformers==4.45.2
networkx==3.4.2
pydantic==2.10.4
python_igraph==0.11.8
tenacity==8.5.0
tiktoken==0.7.0
nest_asyncio
numpy
scipy
tqdm
einops
boto3
PyPDF2
python-docx
atlassian-python-api
Office365-REST-Python-Client
streamlit
pytest
pytest-mock